extracting components from home screen into different files
