+++ app/src/main/res/values/strings.xml
@@ -34,6 +34,7 @@
     <string name="monthly_difference_dialog_text">A diferença é o valor do limite do mês subtraído do valor total das faturas</string>
     <string name="balance_input_hint">Saldo</string>
     <string name="balance_description_input_hint">Descrição</string>
+    <string name="card_bill_input_hint">Fatura</string>
 
     <string name="operations_route">operations</string>
     <string name="balance_title">Balanço</string>
